% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.analisis.R
\name{analize}
\alias{analize}
\title{Analizing a quantitative variable}
\usage{
analize(x)
}
\arguments{
\item{x}{A vector or data.frame with the variables to analize.}
}
\description{
This function shows length, summary, sd and CV of the variable.
}
\examples{
analize()
}
\keyword{analize}
\keyword{summary}
