% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.coefficent.R
\name{CV}
\alias{CV}
\title{Variance coefficent}
\usage{
CV(x)
}
\arguments{
\item{x}{A vector or data frame to calculate the variance coefficent.}
}
\description{
This function calculates the variance coefficient, sd / mean.
}
\examples{
CV()
}
\keyword{coefficient}
