% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambert.beer.law.R
\name{lb.model}
\alias{lb.model}
\title{Lambert-Beer Model Application}
\usage{
lb.model(conc, abs, inter = FALSE)
}
\arguments{
\item{conc}{A vector containing the standard curve concentration.}

\item{abs}{A vector containing the standard curve absorbance.}

\item{inter}{A logical value: if FALSE, intercept is set to 0, otherwise intercept is calculated. Defaults to FALSE.}
}
\description{
Create an Absorbance-Concentration plot
}
\examples{
c -> c(5, 10, 20, 40); a -> c(0.258, 0.489, 0.896, 1.429)
lb.model(c, a, TRUE)

}
\keyword{Lambert-Beer}
\keyword{absorbance}
\keyword{concentration}
