% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality.plots.R
\name{norm.plot}
\alias{norm.plot}
\title{Normal qqplots for more than one variable}
\usage{
norm.plot(x, rows = 1)
}
\arguments{
\item{x}{A data.frame with the variables to plot.}

\item{rows}{The number of rows the plots need to be arranged in. Default = 1}
}
\description{
This function allows you to create many qqnorm plots at once, with qqlines added.
}
\examples{
norm.plot()
}
\keyword{normality}
\keyword{plots}
\keyword{qqline}
\keyword{qqnorm}
\keyword{qqplot}
